+++ server/src-lib/Hasura/RQL/IR/BoolExp.hs
@@ -330,7 +330,9 @@ data ComputedFieldBoolExp (b :: BackendType) a
   = CFBEScalar ![OpExpG b a] -- SQL function returning a scalar
   | CFBETable !(TableName b) !(AnnBoolExp b a) -- SQL function returning SET OF table
   deriving (Functor, Foldable, Traversable, Generic)
-deriving instance (Backend b, Eq (BooleanOperators b a), Eq a) => Eq (ComputedFieldBoolExp b a)
+deriving instance (Backend b, Eq   (BooleanOperators b a), Eq a  ) => Eq   (ComputedFieldBoolExp b a)
+deriving instance (Backend b, Show (BooleanOperators b a), Show a) => Show (ComputedFieldBoolExp b a)
+
 instance (Backend b, NFData    (BooleanOperators b a), NFData    a) => NFData    (ComputedFieldBoolExp b a)
 instance (Backend b, Cacheable (BooleanOperators b a), Cacheable a) => Cacheable (ComputedFieldBoolExp b a)
 instance (Backend b, Hashable  (BooleanOperators b a), Hashable  a) => Hashable  (ComputedFieldBoolExp b a)
@@ -357,7 +359,9 @@ data AnnComputedFieldBoolExp (b :: BackendType) a
   , _acfbSessionArgumentPresence :: !(SessionArgumentPresence a)
   , _acfbBoolExp                 :: !(ComputedFieldBoolExp b a)
   } deriving (Functor, Foldable, Traversable, Generic)
-deriving instance (Backend b, Eq (BooleanOperators b a), Eq a) => Eq (AnnComputedFieldBoolExp b a)
+deriving instance (Backend b, Eq   (BooleanOperators b a), Eq a  ) => Eq   (AnnComputedFieldBoolExp b a)
+deriving instance (Backend b, Show (BooleanOperators b a), Show a) => Show (AnnComputedFieldBoolExp b a)
+
 instance (Backend b, NFData    (BooleanOperators b a), NFData    a) => NFData    (AnnComputedFieldBoolExp b a)
 instance (Backend b, Cacheable (BooleanOperators b a), Cacheable a) => Cacheable (AnnComputedFieldBoolExp b a)
 instance (Backend b, Hashable  (BooleanOperators b a), Hashable  a) => Hashable  (AnnComputedFieldBoolExp b a)
@@ -373,7 +377,9 @@ data AnnBoolExpFld (b :: BackendType) a
   | AVRelationship !(RelInfo b) !(AnnBoolExp b a)
   | AVComputedField !(AnnComputedFieldBoolExp b a)
   deriving (Functor, Foldable, Traversable, Generic)
-deriving instance (Backend b, Eq (BooleanOperators b a), Eq a) => Eq (AnnBoolExpFld b a)
+deriving instance (Backend b, Eq   (BooleanOperators b a), Eq a  ) => Eq   (AnnBoolExpFld b a)
+deriving instance (Backend b, Show (BooleanOperators b a), Show a) => Show (AnnBoolExpFld b a)
+
 instance (Backend b, NFData    (BooleanOperators b a), NFData    a) => NFData    (AnnBoolExpFld b a)
 instance (Backend b, Cacheable (BooleanOperators b a), Cacheable a) => Cacheable (AnnBoolExpFld b a)
 instance (Backend b, Hashable  (BooleanOperators b a), Hashable  a) => Hashable  (AnnBoolExpFld b a)
@@ -477,8 +483,9 @@ $(deriveJSON hasuraJSON ''STIntersectsGeomminNband)
 newtype AnnColumnCaseBoolExpField (b :: BackendType) a
   = AnnColumnCaseBoolExpField { _accColCaseBoolExpField :: AnnBoolExpFld b a }
   deriving (Functor, Foldable, Traversable, Generic)
+deriving instance (Backend b, Eq   (BooleanOperators b a), Eq a  ) => Eq   (AnnColumnCaseBoolExpField b a)
+deriving instance (Backend b, Show (BooleanOperators b a), Show a) => Show (AnnColumnCaseBoolExpField b a)
 
-deriving instance (Backend b, Eq (BooleanOperators b a), Eq a) => Eq (AnnColumnCaseBoolExpField b a)
 instance (Backend b, NFData    (BooleanOperators b a), NFData    a) => NFData    (AnnColumnCaseBoolExpField b a)
 instance (Backend b, Cacheable (BooleanOperators b a), Cacheable a) => Cacheable (AnnColumnCaseBoolExpField b a)
 instance (Backend b, Hashable  (BooleanOperators b a), Hashable  a) => Hashable  (AnnColumnCaseBoolExpField b a)