+++ server/src-lib/Hasura/RQL/Types/Column.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE UndecidableInstances #-}
+
 module Hasura.RQL.Types.Column
   ( ColumnType(..)
   , _ColumnScalar
@@ -113,6 +115,8 @@ data ColumnValue (b :: BackendType) = ColumnValue
   { cvType  :: ColumnType b
   , cvValue :: ScalarValue b
   }
+deriving instance (Backend b, Eq   (ScalarValue b)) => Eq   (ColumnValue b)
+deriving instance (Backend b, Show (ScalarValue b)) => Show (ColumnValue b)
 
 isScalarColumnWhere :: (ScalarType b -> Bool) -> ColumnType b -> Bool
 isScalarColumnWhere f = \case